Move Users to Models dir, and adjust namespaces appropriately in Models Dir.
